pipelines/requirements.txt
PyPDF2
boto3
torch
transformers
opencv-python-headless 
pytesseract
pymupdf
apache-airflow
apache-airflow-providers-amazon
requests